fix(Post.ts): revert namespace change
